Update early access SDK 3.0.9-build20210810 (#45)
