+++ Arduino_package/hardware/boards.txt
@@ -17,7 +17,7 @@ ameba_rtl8721d.build.extra_flags=-mthumb -DRTL8722DM -DBOARD_RTL8722D {build.usb
 ameba_rtl8721d.build.ldscript=linker_scripts/gcc/rlx8721d_img2_is_arduino.ld
 ameba_rtl8721d.build.variant=rtl8721d
 
-ameba_rtl8722dm_mini.name=RTL8722DM_mini
+ameba_rtl8722dm_mini.name=RTL8722DM_MINI
 ameba_rtl8722dm_mini.vid.0=0x2341
 ameba_rtl8722dm_mini.pid.0=0x003d
 ameba_rtl8722dm_mini.upload.tool=ameba_d_tools
