+++ Arduino_package/hardware/platform.txt
@@ -5,7 +5,7 @@
 # For more info:
 # https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5-3rd-party-Hardware-specification
 
-name=Ameba ARM (32-bits) Boards
+name=AmebaD ARM (32-bits) Boards
 version=3.0.9
 
 # Ameba project settings
